backend/pyproject.toml
[tool.poetry]
name = "mestring"
version = "0.0.1"
description = ""
authors = []
readme = "../README.md"


[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
oauthlib = "^3.2.2"
requests = "^2.31.0"
pyopenssl = "^24.0.0"
gunicorn = "^23.0.0"
uvicorn = "^0.27.1"
django = "4.2.20"
djangorestframework = "^3.16.0"
pytest-django = "^4.11.1"
django-cors-headers = "^4.7.0"
nanoid = "^2.0.0"
djangorestframework-camel-case = "^1.4.2"
pyexcel-xlsx = "^0.6.1"
django-filter = "^25.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.28.0"
drf-access-policy = "^1.5.0"

[tool.poetry.group.dev.dependencies]
daphne = "^4.1.0"
poethepoet = "^0.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mastery.tests.settings_in_memory"
python_files = ["test_*.py"]
testpaths = ["mastery/tests"]
django_find_project = true
pythonpath = ["."]


[tool.poetry.group.test.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"


[tool.poe.tasks]
run-tests = "pytest --verbose -s"
run-server = "python manage.py runserver 5000"
